backend/pyproject.toml
[project]
name = "obs-backend"
version = "0.1.0"
description = "Backend for Observability SDK"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.0.1",
    "clickhouse-connect>=0.7.19",
    "bcrypt>=4.2.0",
    "pyjwt>=2.9.0",
    "email-validator>=2.2.0",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.21",
    "greenlet>=3.3.0",
    "observix[eval]",
    "litellm>=1.80.15",
]

[tool.uv.sources]
observix = { path = "observix", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]